+++ MATLAB/vva_read_log.m
@@ -1,7 +1,7 @@
 function [Data, Header] = vva_read_log(fn)
 
 [h, dataString] = klib.file.parse_ini(fn, 'VVA DATA LOG');
-Header = h.Info;
+Header = h.VVA_DATA_LOG;
 
 Data = klib.file.spreadsheet_string_to_struct(dataString);
 Data.Time_s = Data.Time_s - Data.Time_s(1);