+++ VVA Controller/VVA Controller/ControllerSettings.cs
@@ -46,7 +46,7 @@ namespace Jenks.VVA
         {
             get
             {
-                return Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), "Jenks", "VVA Controller Settings.xml");
+                return Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.CommonApplicationData), "Jenks", "VVA Controller Defaults.xml");
             }
         }