+++ VVA Controller/VVA Controller/Properties/Resources.Designer.cs
@@ -73,6 +73,26 @@ namespace VVA_Controller.Properties {
         /// <summary>
         ///   Looks up a localized resource of type System.Drawing.Bitmap.
         /// </summary>
+        internal static System.Drawing.Bitmap dice_blue {
+            get {
+                object obj = ResourceManager.GetObject("dice_blue", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
+        internal static System.Drawing.Bitmap dice_blue1 {
+            get {
+                object obj = ResourceManager.GetObject("dice_blue1", resourceCulture);
+                return ((System.Drawing.Bitmap)(obj));
+            }
+        }
+        
+        /// <summary>
+        ///   Looks up a localized resource of type System.Drawing.Bitmap.
+        /// </summary>
         internal static System.Drawing.Bitmap media_play_green {
             get {
                 object obj = ResourceManager.GetObject("media_play_green", resourceCulture);