+++ VVA Controller/VVA Controller/ScenePropertiesDialog.Designer.cs
@@ -48,6 +48,11 @@
             this.dotSizeNumeric = new KLib.Controls.KNumericBox();
             this.label10 = new System.Windows.Forms.Label();
             this.wallOffsetNumeric = new KLib.Controls.KNumericBox();
+            this.label11 = new System.Windows.Forms.Label();
+            this.amplitudeNumeric = new KLib.Controls.KNumericBox();
+            this.label12 = new System.Windows.Forms.Label();
+            this.durationNumeric = new KLib.Controls.KNumericBox();
+            this.label13 = new System.Windows.Forms.Label();
             this.SuspendLayout();
             // 
             // label1
@@ -55,7 +60,7 @@
             this.label1.AutoSize = true;
             this.label1.Font = new System.Drawing.Font("Microsoft Sans Serif", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label1.ForeColor = System.Drawing.Color.Teal;
-            this.label1.Location = new System.Drawing.Point(341, 18);
+            this.label1.Location = new System.Drawing.Point(124, 190);
             this.label1.Name = "label1";
             this.label1.Size = new System.Drawing.Size(105, 20);
             this.label1.TabIndex = 1;
@@ -64,7 +69,7 @@
             // label2
             // 
             this.label2.AutoSize = true;
-            this.label2.Location = new System.Drawing.Point(251, 53);
+            this.label2.Location = new System.Drawing.Point(34, 222);
             this.label2.Name = "label2";
             this.label2.Size = new System.Drawing.Size(92, 13);
             this.label2.TabIndex = 2;
@@ -73,7 +78,7 @@
             // label3
             // 
             this.label3.AutoSize = true;
-            this.label3.Location = new System.Drawing.Point(243, 79);
+            this.label3.Location = new System.Drawing.Point(26, 248);
             this.label3.Name = "label3";
             this.label3.Size = new System.Drawing.Size(100, 13);
             this.label3.TabIndex = 4;
@@ -82,7 +87,7 @@
             // label4
             // 
             this.label4.AutoSize = true;
-            this.label4.Location = new System.Drawing.Point(273, 104);
+            this.label4.Location = new System.Drawing.Point(56, 273);
             this.label4.Name = "label4";
             this.label4.Size = new System.Drawing.Size(70, 13);
             this.label4.TabIndex = 6;
@@ -92,7 +97,7 @@
             // 
             this.cancelButton.Image = global::VVA_Controller.Properties.Resources.delete;
             this.cancelButton.ImageAlign = System.Drawing.ContentAlignment.MiddleLeft;
-            this.cancelButton.Location = new System.Drawing.Point(396, 188);
+            this.cancelButton.Location = new System.Drawing.Point(148, 430);
             this.cancelButton.Name = "cancelButton";
             this.cancelButton.Padding = new System.Windows.Forms.Padding(10, 0, 10, 0);
             this.cancelButton.Size = new System.Drawing.Size(100, 38);
@@ -106,7 +111,7 @@
             // 
             this.applyButton.Image = global::VVA_Controller.Properties.Resources.ok;
             this.applyButton.ImageAlign = System.Drawing.ContentAlignment.MiddleLeft;
-            this.applyButton.Location = new System.Drawing.Point(290, 188);
+            this.applyButton.Location = new System.Drawing.Point(30, 430);
             this.applyButton.Name = "applyButton";
             this.applyButton.Padding = new System.Windows.Forms.Padding(10, 0, 10, 0);
             this.applyButton.Size = new System.Drawing.Size(100, 38);
@@ -121,7 +126,7 @@
             this.label5.AutoSize = true;
             this.label5.Font = new System.Drawing.Font("Microsoft Sans Serif", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
             this.label5.ForeColor = System.Drawing.Color.Teal;
-            this.label5.Location = new System.Drawing.Point(127, 19);
+            this.label5.Location = new System.Drawing.Point(154, 15);
             this.label5.Name = "label5";
             this.label5.Size = new System.Drawing.Size(39, 20);
             this.label5.TabIndex = 9;
@@ -130,7 +135,7 @@
             // label6
             // 
             this.label6.AutoSize = true;
-            this.label6.Location = new System.Drawing.Point(19, 76);
+            this.label6.Location = new System.Drawing.Point(46, 72);
             this.label6.Name = "label6";
             this.label6.Size = new System.Drawing.Size(80, 13);
             this.label6.TabIndex = 13;
@@ -139,7 +144,7 @@
             // label7
             // 
             this.label7.AutoSize = true;
-            this.label7.Location = new System.Drawing.Point(31, 50);
+            this.label7.Location = new System.Drawing.Point(58, 46);
             this.label7.Name = "label7";
             this.label7.Size = new System.Drawing.Size(68, 13);
             this.label7.TabIndex = 11;
@@ -148,7 +153,7 @@
             // label8
             // 
             this.label8.AutoSize = true;
-            this.label8.Location = new System.Drawing.Point(22, 128);
+            this.label8.Location = new System.Drawing.Point(49, 124);
             this.label8.Name = "label8";
             this.label8.Size = new System.Drawing.Size(77, 13);
             this.label8.TabIndex = 17;
@@ -157,7 +162,7 @@
             // label9
             // 
             this.label9.AutoSize = true;
-            this.label9.Location = new System.Drawing.Point(11, 102);
+            this.label9.Location = new System.Drawing.Point(38, 98);
             this.label9.Name = "label9";
             this.label9.Size = new System.Drawing.Size(88, 13);
             this.label9.TabIndex = 15;
@@ -172,7 +177,7 @@
             this.wallHeightNumeric.FloatValue = 0F;
             this.wallHeightNumeric.IntValue = 0;
             this.wallHeightNumeric.IsInteger = false;
-            this.wallHeightNumeric.Location = new System.Drawing.Point(101, 125);
+            this.wallHeightNumeric.Location = new System.Drawing.Point(128, 121);
             this.wallHeightNumeric.MaxCoerce = false;
             this.wallHeightNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.wallHeightNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -196,7 +201,7 @@
             this.wallDistanceNumeric.FloatValue = 0F;
             this.wallDistanceNumeric.IntValue = 0;
             this.wallDistanceNumeric.IsInteger = false;
-            this.wallDistanceNumeric.Location = new System.Drawing.Point(101, 99);
+            this.wallDistanceNumeric.Location = new System.Drawing.Point(128, 95);
             this.wallDistanceNumeric.MaxCoerce = false;
             this.wallDistanceNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.wallDistanceNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -220,7 +225,7 @@
             this.barSpacingNumeric.FloatValue = 0F;
             this.barSpacingNumeric.IntValue = 0;
             this.barSpacingNumeric.IsInteger = false;
-            this.barSpacingNumeric.Location = new System.Drawing.Point(101, 73);
+            this.barSpacingNumeric.Location = new System.Drawing.Point(128, 69);
             this.barSpacingNumeric.MaxCoerce = false;
             this.barSpacingNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.barSpacingNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -244,7 +249,7 @@
             this.barWidthNumeric.FloatValue = 0F;
             this.barWidthNumeric.IntValue = 0;
             this.barWidthNumeric.IsInteger = false;
-            this.barWidthNumeric.Location = new System.Drawing.Point(101, 47);
+            this.barWidthNumeric.Location = new System.Drawing.Point(128, 43);
             this.barWidthNumeric.MaxCoerce = false;
             this.barWidthNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.barWidthNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -268,7 +273,7 @@
             this.dotVelocityNumeric.FloatValue = 0F;
             this.dotVelocityNumeric.IntValue = 0;
             this.dotVelocityNumeric.IsInteger = false;
-            this.dotVelocityNumeric.Location = new System.Drawing.Point(346, 102);
+            this.dotVelocityNumeric.Location = new System.Drawing.Point(129, 271);
             this.dotVelocityNumeric.MaxCoerce = false;
             this.dotVelocityNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.dotVelocityNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -292,7 +297,7 @@
             this.dotDensityNumeric.FloatValue = 0F;
             this.dotDensityNumeric.IntValue = 0;
             this.dotDensityNumeric.IsInteger = false;
-            this.dotDensityNumeric.Location = new System.Drawing.Point(346, 76);
+            this.dotDensityNumeric.Location = new System.Drawing.Point(129, 245);
             this.dotDensityNumeric.MaxCoerce = false;
             this.dotDensityNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.dotDensityNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -316,7 +321,7 @@
             this.dotSizeNumeric.FloatValue = 0F;
             this.dotSizeNumeric.IntValue = 0;
             this.dotSizeNumeric.IsInteger = false;
-            this.dotSizeNumeric.Location = new System.Drawing.Point(346, 50);
+            this.dotSizeNumeric.Location = new System.Drawing.Point(128, 219);
             this.dotSizeNumeric.MaxCoerce = false;
             this.dotSizeNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.dotSizeNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -334,7 +339,7 @@
             // label10
             // 
             this.label10.AutoSize = true;
-            this.label10.Location = new System.Drawing.Point(25, 154);
+            this.label10.Location = new System.Drawing.Point(52, 150);
             this.label10.Name = "label10";
             this.label10.Size = new System.Drawing.Size(74, 13);
             this.label10.TabIndex = 19;
@@ -349,7 +354,7 @@
             this.wallOffsetNumeric.FloatValue = 0F;
             this.wallOffsetNumeric.IntValue = 0;
             this.wallOffsetNumeric.IsInteger = false;
-            this.wallOffsetNumeric.Location = new System.Drawing.Point(101, 151);
+            this.wallOffsetNumeric.Location = new System.Drawing.Point(128, 147);
             this.wallOffsetNumeric.MaxCoerce = false;
             this.wallOffsetNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
             this.wallOffsetNumeric.MaxValue = 1.7976931348623157E+308D;
@@ -364,11 +369,93 @@
             this.wallOffsetNumeric.Units = "";
             this.wallOffsetNumeric.Value = 0D;
             // 
+            // label11
+            // 
+            this.label11.AutoSize = true;
+            this.label11.Location = new System.Drawing.Point(27, 342);
+            this.label11.Name = "label11";
+            this.label11.Size = new System.Drawing.Size(100, 13);
+            this.label11.TabIndex = 21;
+            this.label11.Text = "Amplitude (degrees)";
+            // 
+            // amplitudeNumeric
+            // 
+            this.amplitudeNumeric.AllowInf = false;
+            this.amplitudeNumeric.AutoSize = true;
+            this.amplitudeNumeric.AutoSizeMode = System.Windows.Forms.AutoSizeMode.GrowAndShrink;
+            this.amplitudeNumeric.ClearOnDisable = false;
+            this.amplitudeNumeric.FloatValue = 0F;
+            this.amplitudeNumeric.IntValue = 0;
+            this.amplitudeNumeric.IsInteger = false;
+            this.amplitudeNumeric.Location = new System.Drawing.Point(129, 339);
+            this.amplitudeNumeric.MaxCoerce = false;
+            this.amplitudeNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
+            this.amplitudeNumeric.MaxValue = 1.7976931348623157E+308D;
+            this.amplitudeNumeric.MinCoerce = false;
+            this.amplitudeNumeric.MinimumSize = new System.Drawing.Size(10, 20);
+            this.amplitudeNumeric.MinValue = 0D;
+            this.amplitudeNumeric.Name = "amplitudeNumeric";
+            this.amplitudeNumeric.Size = new System.Drawing.Size(100, 20);
+            this.amplitudeNumeric.TabIndex = 20;
+            this.amplitudeNumeric.TextFormat = "K4";
+            this.amplitudeNumeric.ToolTip = "";
+            this.amplitudeNumeric.Units = "";
+            this.amplitudeNumeric.Value = 0D;
+            // 
+            // label12
+            // 
+            this.label12.AutoSize = true;
+            this.label12.Location = new System.Drawing.Point(26, 368);
+            this.label12.Name = "label12";
+            this.label12.Size = new System.Drawing.Size(102, 13);
+            this.label12.TabIndex = 23;
+            this.label12.Text = "Baseline duration (s)";
+            // 
+            // durationNumeric
+            // 
+            this.durationNumeric.AllowInf = false;
+            this.durationNumeric.AutoSize = true;
+            this.durationNumeric.AutoSizeMode = System.Windows.Forms.AutoSizeMode.GrowAndShrink;
+            this.durationNumeric.ClearOnDisable = false;
+            this.durationNumeric.FloatValue = 0F;
+            this.durationNumeric.IntValue = 0;
+            this.durationNumeric.IsInteger = false;
+            this.durationNumeric.Location = new System.Drawing.Point(129, 365);
+            this.durationNumeric.MaxCoerce = false;
+            this.durationNumeric.MaximumSize = new System.Drawing.Size(20000, 20);
+            this.durationNumeric.MaxValue = 1.7976931348623157E+308D;
+            this.durationNumeric.MinCoerce = false;
+            this.durationNumeric.MinimumSize = new System.Drawing.Size(10, 20);
+            this.durationNumeric.MinValue = 0D;
+            this.durationNumeric.Name = "durationNumeric";
+            this.durationNumeric.Size = new System.Drawing.Size(100, 20);
+            this.durationNumeric.TabIndex = 22;
+            this.durationNumeric.TextFormat = "K4";
+            this.durationNumeric.ToolTip = "";
+            this.durationNumeric.Units = "";
+            this.durationNumeric.Value = 0D;
+            // 
+            // label13
+            // 
+            this.label13.AutoSize = true;
+            this.label13.Font = new System.Drawing.Font("Microsoft Sans Serif", 12F, System.Drawing.FontStyle.Regular, System.Drawing.GraphicsUnit.Point, ((byte)(0)));
+            this.label13.ForeColor = System.Drawing.Color.Teal;
+            this.label13.Location = new System.Drawing.Point(144, 316);
+            this.label13.Name = "label13";
+            this.label13.Size = new System.Drawing.Size(69, 20);
+            this.label13.TabIndex = 24;
+            this.label13.Text = "Defaults";
+            // 
             // ScenePropertiesDialog
             // 
             this.AutoScaleDimensions = new System.Drawing.SizeF(6F, 13F);
             this.AutoScaleMode = System.Windows.Forms.AutoScaleMode.Font;
-            this.ClientSize = new System.Drawing.Size(531, 279);
+            this.ClientSize = new System.Drawing.Size(275, 501);
+            this.Controls.Add(this.label13);
+            this.Controls.Add(this.label12);
+            this.Controls.Add(this.durationNumeric);
+            this.Controls.Add(this.label11);
+            this.Controls.Add(this.amplitudeNumeric);
             this.Controls.Add(this.label10);
             this.Controls.Add(this.wallOffsetNumeric);
             this.Controls.Add(this.label8);
@@ -420,5 +507,10 @@
         private KLib.Controls.KNumericBox wallDistanceNumeric;
         private System.Windows.Forms.Label label10;
         private KLib.Controls.KNumericBox wallOffsetNumeric;
+        private System.Windows.Forms.Label label11;
+        private KLib.Controls.KNumericBox amplitudeNumeric;
+        private System.Windows.Forms.Label label12;
+        private KLib.Controls.KNumericBox durationNumeric;
+        private System.Windows.Forms.Label label13;
     }
 }