+++ CHANGELOG.md
@@ -1,8 +1,11 @@
 ## Changelog
 
-### Unreleased
+### v1.1 (unreleased)
 #### Changed
 - Discovery: hardwire server to "localhost" to avoid the unnecessary delay of querying Windows for it
+- single test now consists of a static baseline (black, white, dots or bars) followed by moving bars
+#### Added
+- creates a list of all possible tests and randomizes the order
 
 ---